+++ speed_by_Billy_Idol.c
@@ -30,8 +30,8 @@ int get_julia_image(t_julia julia, int *image) {
 
 	//create input data array
 	const int LIST_SIZE = 7;
-	cl_float *A = (cl_float*)malloc(sizeof(cl_float)*LIST_SIZE);
-	A[0] = 4.0 * julia.zoom/SIZE; // float increment
+	cl_double *A = (cl_double*)malloc(sizeof(cl_double)*LIST_SIZE);
+	A[0] = 4.0 * julia.zoom/SIZE; // double increment
 	A[1] = julia.c.re; // ReC
 	A[2] = julia.c.im; //ImC
 	A[3] = julia.acc; //Max iter, will be converted to int
@@ -50,11 +50,11 @@ int get_julia_image(t_julia julia, int *image) {
 	cl_context context = clCreateContext(NULL, 1, &device_id, NULL, NULL, &ret);
 	cl_command_queue command_queue = clCreateCommandQueue(context, device_id, 0, &ret);
 	cl_mem a_mem_obj = clCreateBuffer(context, CL_MEM_READ_ONLY,
-									  LIST_SIZE * sizeof(cl_float), NULL, &ret);
+									  LIST_SIZE * sizeof(cl_double), NULL, &ret);
 	cl_mem c_mem_obj = clCreateBuffer(context, CL_MEM_READ_WRITE,
 									  SIZE * SIZE * sizeof(cl_int), NULL, &ret);
 	ret = clEnqueueWriteBuffer(command_queue, a_mem_obj, CL_TRUE, 0,
-							   LIST_SIZE * sizeof(cl_float), A, 0, NULL, NULL);
+							   LIST_SIZE * sizeof(cl_double), A, 0, NULL, NULL);
 	cl_program program = clCreateProgramWithSource(context, 1,
 												   (const char **) &source, (const size_t *) &f_size, &ret);
 	ret = clBuildProgram(program, 1, &device_id, NULL, NULL, NULL);
